+++ src/app.jsx
@@ -1,33 +1,33 @@
-import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import { SnackbarProvider } from "notistack";
+import React from 'react';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { SnackbarProvider } from 'notistack';
 
-import GlobalStyles from "./GlobalStyles";
+import GlobalStyles from './GlobalStyles';
 
-import Home from "./pages/Home";
-import Detail from "./pages/Detail";
-import RegisterPage from "./components/register/RegisterPage";
-import MyPage from "./pages/MyPage";
+import Home from './pages/Home';
+import Detail from './pages/Detail';
+import Register from './pages/Register';
+import MyPage from './pages/MyPage';
 
 const App = () => {
-	return (
-		<BrowserRouter>
-			<SnackbarProvider
-				anchorOrigin={{
-					vertical: "top",
-					horizontal: "right",
-				}}
-			>
-				<GlobalStyles />
-				<Routes>
-					<Route path="/" element={<Home />} />
-					<Route path="/detail/:id" element={<Detail />} />
-					<Route path="/register" element={<RegisterPage />} />
-					<Route path="/my" element={<MyPage />} />
-				</Routes>
-			</SnackbarProvider>
-		</BrowserRouter>
-	);
+  return (
+    <BrowserRouter>
+      <SnackbarProvider
+        anchorOrigin={{
+          vertical: 'top',
+          horizontal: 'right',
+        }}
+      >
+        <GlobalStyles />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/detail/:id" element={<Detail />} />
+          <Route path="/register" element={<Register />} />
+          <Route path="/my" element={<MyPage />} />
+        </Routes>
+      </SnackbarProvider>
+    </BrowserRouter>
+  );
 };
 
 export default App;
