+++ src/pages/Register.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import Navbar from '../components/common/Navbar';
+import Main from '../components/register/Main';
+import Footer from '../components/common/Footer';
+
+const Register = () => {
+  return (
+    <>
+      <Navbar />
+      <Main />
+      <Footer />
+    </>
+  );
+};
+
+export default Register;
